OAK-11625: Move o.a.j.o.plugins.bundlor.BundlingConfigInitializer to a suitable API package and export that package.
Renamed package to avoid confusion.
